add docs updates, merge weather skills, fix issues detected
