+++ aea/cli/add.py
@@ -64,7 +64,7 @@ def _find_connection_locally(ctx, connection_public_id, click_context):
     try:
         connection_configuration = ctx.connection_loader.load(open(str(connection_configuration_filepath)))
         if connection_configuration.restricted_to_protocols != set():
-            logger.info("Connection '{}' is restricted to the following protocols: {}".format(connection_name, connection_configuration.restricted_to_protocols))
+            logger.info("Connection '{}' is restricted to the following protocols: {}".format(connection_name, [str(protocol_id) for protocol_id in connection_configuration.restricted_to_protocols]))
     except ValidationError as e:
         logger.error("Connection configuration file not valid: {}".format(str(e)))
         sys.exit(1)
