+++ docs/generator.md
@@ -1,0 +1,5 @@
+<div class="admonition note">
+  <p class="admonition-title">Note</p>
+  <p>Coming soon!</p>
+</div>
+
