+++ docs/gym-skill.md
@@ -1,25 +1,25 @@
-The AEA gym skill demonstrates how a custom Reinforcement Learning agent, that uses OpenAI's <a href="https://gym.openai.com" target=_blank>gym</a> library, may be embedded into an Autonomous Economic Agent.
+The AEA gym skill demonstrates how a custom Reinforcement Learning agent, that uses OpenAI's <a href="https://gym.openai.com" target=_blank>gym</a> library, may be embedded into an AEA skill and connection.
 
 
-## Demo instructions
+## Preparation instructions
 
 ### Dependencies
 
 Follow the <a href="../quickstart/#preliminaries">Preliminaries</a> and <a href="../quickstart/#installation">Installation</a> sections from the AEA quick start.
 
+## Demo instructions
+
 ### Create the agent
-In the root directory, create the gym agent.
+In the root directory, create the gym agent and enter the project.
 ``` bash
 aea create my_gym_agent
+cd my_gym_agent
 ```
-
 
 ### Add the gym skill 
 ``` bash
-cd my_gym_agent
 aea add skill fetchai/gym:0.1.0
 ```
-
 
 ### Copy the gym environment to the agent directory
 ``` bash
@@ -27,12 +27,10 @@ mkdir gyms
 cp -a ../examples/gym_ex/gyms/. gyms/
 ```
 
-
 ### Add a gym connection
 ``` bash
 aea add connection fetchai/gym:0.1.0
 ```
-
 
 ### Update the connection config
 ``` bash
@@ -67,7 +65,7 @@ cd ..
 aea delete my_gym_agent
 ```
 
-### Communication
+## Communication
 This diagram shows the communication between the agent and the gym environment 
 
 <div class="mermaid">
