+++ docs/index.md
@@ -2,13 +2,13 @@
 
 The AEA framework provides the tools for creating Autonomous Economic Agents (AEA). AEAs work continuously for your benefit without you having to do anything more than writing them and starting them up.
 
-AEAs act independently of constant input and autonomously develop new capabilities by acquiring new skills. Their goal is to create economic value for you, their owner. AEAs have a wide range of application areas. Check out the <a href="../car-park" target=_blank>Demos section</a> from the left menu for some examples.
+AEAs act independently of constant user input and autonomously execute actions to achieve their objective. Their goal is to create economic value for you, their owner. AEAs have a wide range of application areas. Check out the <a href="../app-areas">Application Areas section</a>  and <a href="../car-park-skills">Demos section</a> for some examples.
 
 The AEA framework is a Python-based development suite which equips you with an efficient and easy to understand set of tools for building autonomous economic agents. The framework is truly modular, easily extensible, and highly composable. This framework attempts to make agent development as straightforward an experience as possible, similar to web development using popular web frameworks.
 
 To start learning about some of the distinctive characteristics of agent-oriented development, check out the <a href="../agent-oriented-development">next page</a>.
 
-AEAs achieve their goals with the help of the OEF - a search and discovery platform for agents by Fetch.ai - and using Fetch.ai's blockchain as a financial system. Third-party systems, such as Ethereum, may also allow AEA integration.
+AEAs achieve their goals with the help of the OEF - a search and discovery platform for agents by Fetch.ai - and using Fetch.ai's blockchain as a financial settlement layer. Third-party blockchains, such as Ethereum, may also allow AEA integration.
 
 <div class="admonition note">
   <p class="admonition-title">Note</p>
