+++ docs/ml-skills.md
@@ -5,12 +5,14 @@ There are two types of agents:
 * The ml data provider which sells training data.
 * The ml model trainer which trains a model
 
+## Preparation instructions
+
 ### Dependencies
 
 Follow the <a href="../quickstart/#preliminaries">Preliminaries</a> and <a href="../quickstart/#installation">Installation</a> sections from the AEA quick start.
 
 
-## Launch an OEF node
+### Launch an OEF node
 In a separate terminal, launch a local OEF node (for search and discovery).
 ``` bash
 python scripts/oef/launch.py -c ./scripts/oef/launch_config.json
@@ -18,26 +20,29 @@ python scripts/oef/launch.py -c ./scripts/oef/launch_config.json
 
 Keep it running for all the following demos.
 
-## Demo 1 - no ledger:
+## Demo instructions 1 - no ledger payment:
 
 
 ### Create the data provider AEA
-In the root directory, create the data provider AEA.
+In the root directory, create the data provider AEA and enter the project.
 ``` bash
 aea create ml_data_provider
+cd ml_data_provider
 ```
 
-### Add the `ml_data_provider` skill
+Add the `ml_data_provider` skill and `oef` connection.
 ``` bash
-cd ml_data_provider
 aea add connection fetchai/oef:0.1.0
 aea add skill fetchai/ml_data_provider:0.1.0
 ```
 
 ### Alternatively, install the agent directly
+In the root directory, fetch the data provider AEA and enter the project.
 ``` bash
 aea fetch fetchai/ml_data_provider:0.1.0
+cd ml_data_provider
 ```
+The `aea fetch` command creates the entire agent, including its dependencies for you.
 
 ### Install the dependencies
 The ml data provider uses `tensorflow` and `numpy`.
@@ -51,21 +56,23 @@ aea run --connections oef
 ```
 
 ### Create the model trainer AEA
-In a separate terminal, in the root directory, create the model trainer AEA.
+In a separate terminal, in the root directory, create the model trainer AEA and enter the project.
 ``` bash
 aea create ml_model_trainer
+cd ml_model_trainer
 ```
 
-### Add the `ml_train` skill to the model trainer AEA
+Add the `ml_train` skill and `oef` connection.
 ``` bash
-cd ml_model_trainer
 aea add connection fetchai/oef:0.1.0
 aea add skill fetchai/ml_train:0.1.0
 ```
 
 ### Alternatively, install the agent directly
+In the root directory, fetch the data provider AEA and enter the project.
 ``` bash
 aea fetch fetchai/ml_model_trainer:0.1.0
+cd ml_model_trainer
 ```
 
 ### Install the dependencies
@@ -82,10 +89,10 @@ aea run --connections oef
 After some time, you should see the AEAs transact and the model trainer train its model.
 
 
-## Demo 2 - ledger:
+## Demo instructions 2 - Fetch.ai ledger payment:
 
 
-We will now run the same demo but with a real ledger transaction on test net.
+We will now run the same demo but with a real ledger transaction on Fetch.ai test net.
 
 ### Update the AEA configs
 
@@ -129,7 +136,7 @@ aea delete ml_data_provider
 aea delete ml_model_trainer
 ```
 
-### Communication
+## Communication
 This diagram shows the communication between the two agents.
 
 <div class="mermaid">
