+++ docs/skill-guide.md
@@ -1,5 +1,3 @@
-The <a href="../scaffolding/">scaffolding tool</a> allows you to create the folder structure required for a skill.
-
 <div class="admonition note">
   <p class="admonition-title">Note</p>
   <p>Before developing your first skill, please read the <a href="../skill/">skill guide</a>.</p>
