+++ docs/tac-skills.md
@@ -5,11 +5,13 @@ There are two types of agents:
 * The tac controller which coordinates the game.
 * The participant agents which compete in the game.
 
+## Preparation instructions
+
 ### Dependencies
 
 Follow the <a href="../quickstart/#preliminaries">Preliminaries</a> and <a href="../quickstart/#installation">Installation</a> sections from the AEA quick start.
 
-## Launch an OEF node
+### Launch an OEF node
 In a separate terminal, launch a local OEF node (for search and discovery).
 ``` bash
 python scripts/oef/launch.py -c ./scripts/oef/launch_config.json
@@ -17,7 +19,7 @@ python scripts/oef/launch.py -c ./scripts/oef/launch_config.json
 
 Keep it running for all the following demos.
 
-## Demo 1: no ledger transactions
+## Demo instructions 1: no ledger transactions
 
 This demo uses another agent - a controller agent - to take the role of running the competition and validating the transactions negotiated by the agents. 
 
@@ -105,6 +107,12 @@ aea config set agent.default_ledger ethereum
 aea run --connections oef
 ```
 	
+## Communication
+
+There are two types of interactions:
+- between the participants and the controller, the game communication
+- between the participants, the negotiation
+
 ### Registration communication
 This diagram shows the communication between the various entities during the registration phase. 
 
@@ -241,19 +249,19 @@ The `OEFSearchHandler` deals with `OEFMessage` types returned from the OEF searc
 
 The `TransactionCleanUpTask` is responsible for cleaning up transactions which are no longer likely to being settled with the controller agent.
 
-## Shared classes
+### Shared classes
 
 The `shared_classes` element in the configuration `yaml` lists a number of important classes which are shared between the handlers, behaviours and tasks.
 
-### Search
+#### Search
 
 This class abstracts the logic required by agents performing searches for other buying/selling agents according to strategy (see below).
 
-### Registration
+#### Registration
 
 This class abstracts the logic required by agents performing service registrations on the OEF.
 
-### Strategy
+#### Strategy
 
 This class defines the strategy behind an agent's activities.
 
@@ -261,10 +269,10 @@ The class is instantiated with the agent's goals, for example whether the agent 
 
 It also provides methods for defining what goods agents are looking for and what goods they may have to sell, for generating proposal queries, and checking whether a proposal is profitable or not.
 
-### Dialogue
+#### Dialogue
 
 `Dialogues` abstract the negotiations that take place between agents including all negotiation end states, such as accepted, declined, etc. and all the negotiation states in between.
 
-### Transactions
+#### Transactions
 
 This class deals with representing potential transactions between agents.
