+++ packages/skills/weather_client/skill.yaml
@@ -16,6 +16,9 @@ handlers:
   oef:
     class_name: OEFHandler
     args: {}
+  transaction:
+    class_name: MyTransactionHandler
+    args: {}
 tasks: {}
 shared_classes:
   strategy:
@@ -25,9 +28,10 @@ shared_classes:
       max_row_price: 4
       max_buyer_tx_fee: 1
       currency_id: 'FET'
-      ledger_id: 'None'
+      ledger_id: 'fetchai'
+      is_ledger_tx: False
   dialogues:
     class_name: Dialogues
     args: {}
 protocols: ['fetchai/fipa:0.1.0','fetchai/default:0.1.0','fetchai/oef:0.1.0']
-ledgers: []
+ledgers: ['fetchai']
