--- packages/skills/weather_client_ledger/skill.yaml
@@ -1,36 +1,0 @@
-name: weather_client_ledger
-author: fetchai
-version: 0.1.0
-license: Apache 2.0
-fingerprint: ""
-description: "The weather client skill implements the skill to purchase weather data."
-behaviours:
-  search:
-    class_name: MySearchBehaviour
-    args:
-      search_interval: 5
-handlers:
-  fipa:
-    class_name: FIPAHandler
-    args: {}
-  oef:
-    class_name: OEFHandler
-    args: {}
-  transaction:
-    class_name: MyTransactionHandler
-    args: {}
-tasks: {}
-shared_classes:
-  strategy:
-    class_name: Strategy
-    args:
-      country: UK
-      max_row_price: 4
-      max_buyer_tx_fee: 1
-      currency_id: 'FET'
-      ledger_id: 'fetchai'
-  dialogues:
-    class_name: Dialogues
-    args: {}
-protocols: ['fetchai/fipa:0.1.0','fetchai/default:0.1.0','fetchai/oef:0.1.0']
-ledgers: ['fetchai']
