--- packages/skills/weather_station_ledger/dialogues.py
@@ -1,63 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# ------------------------------------------------------------------------------
-#
-#   Copyright 2018-2019 Fetch.AI Limited
-#
-#   Licensed under the Apache License, Version 2.0 (the "License");
-#   you may not use this file except in compliance with the License.
-#   You may obtain a copy of the License at
-#
-#       http://www.apache.org/licenses/LICENSE-2.0
-#
-#   Unless required by applicable law or agreed to in writing, software
-#   distributed under the License is distributed on an "AS IS" BASIS,
-#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#   See the License for the specific language governing permissions and
-#   limitations under the License.
-#
-# ------------------------------------------------------------------------------
-
-"""
-This module contains the classes required for dialogue management.
-
-- Dialogue: The dialogue class maintains state of a dialogue and manages it.
-- Dialogues: The dialogues class keeps track of all dialogues.
-"""
-
-from typing import Any, Dict, Optional
-
-from aea.helpers.dialogue.base import DialogueLabel
-from aea.helpers.search.models import Description
-from aea.skills.base import SharedClass
-from packages.protocols.fipa.dialogues import FIPADialogues, FIPADialogue
-
-
-class Dialogue(FIPADialogue):
-    """The dialogue class maintains state of a dialogue and manages it."""
-
-    def __init__(self, dialogue_label: DialogueLabel, is_seller: bool) -> None:
-        """
-        Initialize a dialogue label.
-
-        :param dialogue_label: the identifier of the dialogue
-        :param is_seller: indicates whether the agent associated with the dialogue is a seller or buyer
-
-        :return: None
-        """
-        FIPADialogue.__init__(self, dialogue_label=dialogue_label, is_seller=is_seller)
-        self.weather_data = None  # type: Optional[Dict[str, Any]]
-        self.proposal = None  # type: Optional[Description]
-
-
-class Dialogues(SharedClass, FIPADialogues):
-    """The dialogues class keeps track of all dialogues."""
-
-    def __init__(self, **kwargs) -> None:
-        """
-        Initialize dialogues.
-
-        :return: None
-        """
-        SharedClass.__init__(self, **kwargs)
-        FIPADialogues.__init__(self)
